V2/tests/SampleApps/python/mysql-mysqlconnector-sample/requirements.txt
Flask==2.2.5
mysql-connector==2.6.0
